Full Component refactor and util functions added
