+++ frontend/vite-project/adminContainers/TeamsContainer.jsx
@@ -5,10 +5,11 @@ import { use, useEffect, useState } from "react";
 import axios from "axios";
 import InputFields from "../components/InputFields";
 import Buttons from "../components/Buttons";
+import convertDate from "../utilities/convertDate";
+import TeamsCard from "../components/TeamsCard";
 
 const TeamsContainer = () => {
   const [date, setDate] = useState(null);
-  const [employees, setEmployees] = useState([]);
   const [teams, setTeams] = useState({});
   const [bonus, setBonus] = useState({});
 
@@ -17,8 +18,6 @@ const TeamsContainer = () => {
 
     axios.get(`http://localhost:8081/lass/teams/${fullDate}`).then((res) => {
       const data = res.data.data;
-      setEmployees(data);
-
       const teamsObj = {};
       if (data) {
         data.forEach((employee) => {
@@ -26,11 +25,9 @@ const TeamsContainer = () => {
           if (!teamsObj[teamID]) teamsObj[teamID] = [];
           teamsObj[teamID].push(employee);
         });
-        setEmployees(data);
         setTeams(teamsObj);
       } else {
         setTeams({});
-        setEmployees([]);
       }
     });
   };
@@ -59,184 +56,26 @@ const TeamsContainer = () => {
       <DateSelector
         onChange={(value) => {
           const dateValue = value.$d;
-          const dateStr = new Date(dateValue.toString());
-          const year = dateStr.getFullYear();
-          const month = String(dateStr.getMonth() + 1).padStart(2, "0");
-          const day = String(dateStr.getDate()).padStart(2, "0");
-          const fullDate = `${year}-${month}-${day}`;
-          fetchTeamsByDate(fullDate);
-          setDate(fullDate);
+          const convertedDate = convertDate(dateValue);
+          fetchTeamsByDate(convertedDate);
+          setDate(convertedDate);
         }}
       />
       {date && Object.keys(teams).length > 0 && (
-        <Box
-          sx={{
-            backgroundColor: Colors.content,
-            height: "fit-content",
-            maxHeight: "800px",
-            width: "1200px",
-            border: `2px solid ${Colors.secondary}`,
-            borderRadius: "20px",
-            padding: "20px",
+        <TeamsCard
+          teams={teams}
+          onChange={(team_id, value) => {
+            setBonus((prev) => ({
+              ...prev,
+              [team_id]: value,
+            }));
           }}
-        >
-          <Box
-            sx={{
-              display: "grid",
-              gridTemplateColumns: "repeat(3,1fr)",
-              alignItems: "end",
-              gap: "20px",
-            }}
-          >
-            {Object.entries(teams).map(([team_id, team]) => (
-              <Box
-                sx={{
-                  width: "100%",
-                  display: "flex",
-                  flexDirection: "column",
-                  alignItems: "center",
-                }}
-              >
-                <Box
-                  key={team_id}
-                  sx={{
-                    display: "flex",
-                    flexDirection: "column",
-                    backgroundColor: Colors.primary,
-                    color: Colors.secondary,
-                    borderRadius: "8px",
-                    padding: "20px 0px",
-                    width: "100%",
-                  }}
-                >
-                  <Box
-                    key={team_id}
-                    sx={{
-                      display: "flex",
-                      flexDirection: "row",
-                      justifyContent: "center",
-                      gap: "5px",
-                      alignItems: "center",
-                      color: Colors.content,
-                    }}
-                  >
-                    {team.length > 0 && (
-                      <>
-                        <Typography sx={{}} variant="h5">
-                          {team[0].shift}:
-                        </Typography>
-                        <Typography sx={{}} variant="h5">
-                          Team {team_id}
-                        </Typography>
-                      </>
-                    )}
-                  </Box>
-
-                  {team.map((employee) => (
-                    <Box
-                      key={employee.employee_id}
-                      sx={{
-                        display: "flex",
-                        flexDirection: "row",
-                        alignItems: "center",
-                        justifyContent: "center",
-                      }}
-                    >
-                      <Typography
-                        sx={{
-                          flex: "1",
-                          padding: "5px",
-                          textAlign: "end",
-                          fontWeight: "bold",
-                          color: "lightgray",
-                        }}
-                      >
-                        {employee.role}:
-                      </Typography>
-                      <Typography
-                        sx={{
-                          flex: "1",
-                          padding: "5px",
-                          textAlign: "start",
-                          fontWeight: "bold",
-                          color: Colors.secondary,
-                        }}
-                      >
-                        {employee.employee_name}
-                      </Typography>
-                    </Box>
-                  ))}
-                  <Box
-                    sx={{
-                      display: "flex",
-                      flexDirection: "row",
-                      alignItems: "center",
-                      justifyContent: "center",
-                    }}
-                  >
-                    <Typography
-                      sx={{
-                        flex: "1",
-                        padding: "5px",
-                        textAlign: "end",
-                        fontWeight: "bold",
-                        color: "lightgray",
-                      }}
-                    >
-                      Bonus:
-                    </Typography>
-                    <Typography
-                      sx={{
-                        flex: "1",
-                        padding: "5px",
-                        textAlign: "start",
-                        fontWeight: "bold",
-                        color: Colors.success,
-                      }}
-                    >
-                      ${team[0].bonus === null ? "0" : team[0].bonus}
-                    </Typography>
-                  </Box>
-                </Box>
-
-                <Box
-                  sx={{
-                    display: "flex",
-                    flexDirection: "row",
-                    gap: "5px",
-                    alignItems: "center",
-                    marginTop: "5px",
-                    width: "100%",
-                  }}
-                >
-                  <InputFields
-                    type={"number"}
-                    label={"Update Bonus..."}
-                    width={"60%"}
-                    value={bonus[team_id] || ""}
-                    margin={"none"}
-                    onChange={(e) => {
-                      setBonus((prev) => ({
-                        ...prev,
-                        [team_id]: e.target.value,
-                      }));
-                    }}
-                  />
-                  <Buttons
-                    content={"Submit"}
-                    onClick={() => {
-                      changeBonus(team_id);
-                      setBonus((prev) => ({ ...prev, [team_id]: "" }));
-                    }}
-                    height="56px"
-                    width="40%"
-                  />
-                </Box>
-              </Box>
-            ))}
-            <Box></Box>
-          </Box>
-        </Box>
+          onClick={(team_id) => {
+            changeBonus(team_id);
+            setBonus((prev) => ({ ...prev, [team_id]: "" }));
+          }}
+          bonus={bonus}
+        />
       )}
       {!date && (
         <Typography sx={{ color: "gray" }} variant="h5">
