+++ frontend/vite-project/utilities/dateToDMY.js
@@ -1,0 +1,9 @@
+const dateToDMY = (date) => {
+  const dateArr = date.split("-");
+  const year = dateArr[0];
+  const month = dateArr[1];
+  const day = dateArr[2];
+  return `${day}-${month}-${year}`;
+};
+
+export default dateToDMY;
